Separate simulator and CR server
