+++ simulator/config.go
@@ -4,7 +4,6 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"net"
 	"time"
 
 	"github.com/rs/zerolog"
@@ -18,6 +17,10 @@ var Config struct {
 	// LogLevel controls how verbose the levels are. Supported values: trace,
 	// debug, info, warn, error, fatal, panic.
 	LogLevel string `env:"LOG_LEVEL, default=info"`
+
+	// ConnReqEnableHTTP enables an HTTP server that can accept connection
+	// requests.
+	ConnReqEnableHTTP bool `env:"CR_HTTP, default=true"`
 
 	// Host is the host name or IP address used by the simulator to accept
 	// connection requests. If no value is provided it will be automatically
@@ -99,13 +102,6 @@ func LoadConfig(ctx context.Context) error {
 		return fmt.Errorf("load env config: %w", err)
 	}
 
-	if Config.Host == "" {
-		Config.Host, err = getIP()
-		if err != nil {
-			return fmt.Errorf("get ip address: %w", err)
-		}
-	}
-
 	if Config.ACSAuth != AuthNone {
 		if Config.ACSUsername == "" || Config.ACSPassword == "" {
 			return fmt.Errorf("auth %s: %w", Config.ACSAuth, ErrNoCreds)
@@ -119,18 +115,4 @@ func LoadConfig(ctx context.Context) error {
 	log.Logger = log.Logger.Level(logLevel)
 
 	return nil
-}
-
-func getIP() (string, error) {
-	addrs, err := net.InterfaceAddrs()
-	if err != nil {
-		return "", err
-	}
-	for _, addr := range addrs {
-		ipNet, ok := addr.(*net.IPNet)
-		if ok && !ipNet.IP.IsLoopback() && ipNet.IP.To4() != nil {
-			return ipNet.IP.String(), nil
-		}
-	}
-	return "0.0.0.0", nil
 }
