+++ simulator/get_parameter_names.go
@@ -6,7 +6,7 @@ import (
 	"github.com/localhots/SimulaTR69/rpc"
 )
 
-func (s *Server) handleGetParameterNames(envID string, r *rpc.GetParameterNamesRequest) *rpc.EnvelopeEncoder {
+func (s *Simulator) handleGetParameterNames(envID string, r *rpc.GetParameterNamesRequest) *rpc.EnvelopeEncoder {
 	log.Info().Str("method", "GetParameterNames").Msg("Received message")
 	r.Debug()
 	names := s.dm.ParameterNames(r.ParameterPath, r.NextLevel)
