+++ simulator/get_rpc_methods.go
@@ -6,7 +6,7 @@ import (
 	"github.com/localhots/SimulaTR69/rpc"
 )
 
-func (s *Server) handleGetRPCMethods(envID string) *rpc.EnvelopeEncoder {
+func (s *Simulator) handleGetRPCMethods(envID string) *rpc.EnvelopeEncoder {
 	log.Info().Str("method", "GetRPCMethods").Msg("Received message")
 	methods := rpc.SupportedMethods()
 	for _, m := range methods {
