+++ simulator/reboot.go
@@ -6,7 +6,7 @@ import (
 	"github.com/localhots/SimulaTR69/rpc"
 )
 
-func (s *Server) handleReboot(envID string, r *rpc.RebootRequest) *rpc.EnvelopeEncoder {
+func (s *Simulator) handleReboot(envID string, r *rpc.RebootRequest) *rpc.EnvelopeEncoder {
 	log.Info().Str("method", "Reboot").Msg("Received message")
 	resp := rpc.NewEnvelope(envID)
 	resp.Body.RebootResponse = &rpc.RebootResponseEncoder{}
