+++ simulator/upload.go
@@ -5,7 +5,7 @@ import (
 	"github.com/rs/zerolog/log"
 )
 
-func (s *Server) handleUpload(envID string, r *rpc.UploadRequest) *rpc.EnvelopeEncoder {
+func (s *Simulator) handleUpload(envID string, r *rpc.UploadRequest) *rpc.EnvelopeEncoder {
 	log.Info().Str("method", "Upload").Msg("Received message")
 	r.Debug()
 	// s.dm.SetCommandKey(r.CommandKey)
